posts/annonce-programme-journee-bluehats-2022.md: Update
Since fd0d487, and because I was using the "gazette" layout for the
latest blog post, it started with an invitation to subscribe to the
BlueHats gazette - this was confusing.

fd0d487 is right and I was wrong to (ab)use the gazette layout.

This commit reverts to using the post layout and update the intro
text.